skyvan/requirements.txt
asgiref==3.8.1
attrs==25.1.0
Brotli==1.1.0
cffi==2.0.0
cssselect2==0.8.0
dj-database-url==2.1.0
Django==4.2
django-cors-headers==4.2.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.27.2
et_xmlfile==2.0.0
fonttools==4.60.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mysql-connector-python==8.0.33
mysqlclient==2.1.1
numpy==2.0.2
openpyxl==3.1.5
packaging==24.0
pandas==2.2.3
pillow==11.1.0
protobuf==3.20.3
pycparser==2.23
pydyf==0.11.0
PyJWT==1.7.1
pyphen==0.17.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
referencing==0.36.2
rpds-py==0.23.1
simplejwt==2.0.1
six==1.16.0
sqlparse==0.4.4
tinycss2==1.4.0
tinyhtml5==2.0.0
typing==3.7.4.3
typing_extensions==4.11.0
tzdata==2025.1
uritemplate==4.1.1
weasyprint==66.0
webencodings==0.5.1
xlrd==2.0.1
zopfli==0.2.3.post1
django-debug-toolbar==4.3.0